WIP optimisation, c'est le bordel grand dieu

+++ testAutocad/ComputePlacementVisitor.h
@@ -8,8 +8,8 @@
 #include <mathfu/vector_3.h>
 #include <mathfu/matrix_4x4.h>
 
-typedef mathfu::Vector<float, 3> Vec3;
-typedef mathfu::Matrix<float, 4> Matrix4;
+typedef mathfu::Vector<double, 3> Vec3;
+typedef mathfu::Matrix<double, 4> Matrix4;
 
 class ComputePlacementVisitor : public ifc2x3::InheritVisitor
 {
